+++ sketch-core/src/desktopMain/kotlin/com/github/panpf/sketch/fetch/KotlinResourceUriFetcher.kt
@@ -2,32 +2,38 @@ package com.github.panpf.sketch.fetch
 
 import com.github.panpf.sketch.Sketch
 import com.github.panpf.sketch.annotation.WorkerThread
-import com.github.panpf.sketch.source.KotlinResourceDataSource
 import com.github.panpf.sketch.request.ImageRequest
+import com.github.panpf.sketch.source.KotlinResourceDataSource
 import com.github.panpf.sketch.util.MimeTypeMap
-import com.github.panpf.sketch.util.ifOrNull
+import com.github.panpf.sketch.util.Uri
+import com.github.panpf.sketch.util.pathSegments
 import com.github.panpf.sketch.util.toUri
 
 /**
- * Sample: 'kotlin.resource://test.png'
+ * Sample: 'file://kotlin_resource/test.png'
  */
 fun newKotlinResourceUri(resourceName: String): String =
-    "${KotlinResourceUriFetcher.SCHEME}://$resourceName"
+    "${KotlinResourceUriFetcher.SCHEME}://${KotlinResourceUriFetcher.AUTHORITY}/$resourceName"
+
+fun isKotlinResourceUri(uri: Uri): Boolean =
+    KotlinResourceUriFetcher.SCHEME.equals(uri.scheme, ignoreCase = true)
+            && KotlinResourceUriFetcher.AUTHORITY.equals(uri.authority, ignoreCase = true)
 
 class KotlinResourceUriFetcher(
     val sketch: Sketch,
     val request: ImageRequest,
-    val resourceName: String,
+    val resourcePath: String,
 ) : Fetcher {
 
     companion object {
-        const val SCHEME = "kotlin.resource"
+        const val SCHEME = "file"
+        const val AUTHORITY = "kotlin_resource"
     }
 
     @WorkerThread
     override suspend fun fetch(): Result<FetchResult> = kotlin.runCatching {
-        val mimeType = MimeTypeMap.getMimeTypeFromUrl(resourceName)
-        val dataSource = KotlinResourceDataSource(sketch, request, resourceName)
+        val mimeType = MimeTypeMap.getMimeTypeFromUrl(resourcePath)
+        val dataSource = KotlinResourceDataSource(sketch, request, resourcePath)
         return Result.success(FetchResult(dataSource, mimeType))
     }
 
@@ -35,9 +41,11 @@ class KotlinResourceUriFetcher(
 
         override fun create(sketch: Sketch, request: ImageRequest): KotlinResourceUriFetcher? {
             val uri = request.uri.toUri()
-            return ifOrNull(SCHEME.equals(uri.scheme, ignoreCase = true)) {
-                val resourcePath = "${uri.authority.orEmpty()}${uri.path.orEmpty()}"
+            return if (isKotlinResourceUri(uri)) {
+                val resourcePath = uri.pathSegments.joinToString("/")
                 KotlinResourceUriFetcher(sketch, request, resourcePath)
+            } else {
+                null
             }
         }