+++ sketch-core/src/iosMain/kotlin/com/github/panpf/sketch/source/KotlinResourceDataSource.kt
@@ -11,21 +11,21 @@ import platform.Foundation.NSBundle
 class KotlinResourceDataSource(
     override val sketch: Sketch,
     override val request: ImageRequest,
-    val resourceName: String,
+    val resourcePath: String,
 ) : DataSource {
 
     override val dataFrom: DataFrom
         get() = LOCAL
 
     override fun openSourceOrNull(): Source {
-        val resourcePath = NSBundle.mainBundle.resourcePath!!.toPath()
-        val filePath = resourcePath.resolve("compose-resources").resolve(resourceName)
+        val appResourcePath = NSBundle.mainBundle.resourcePath!!.toPath()
+        val filePath = appResourcePath.resolve("compose-resources").resolve(resourcePath)
         return sketch.fileSystem.source(filePath)
     }
 
     override fun getFileOrNull(): Path {
         val resourcePath = NSBundle.mainBundle.resourcePath!!.toPath()
-        return resourcePath.resolve("compose-resources").resolve(resourceName)
+        return resourcePath.resolve("compose-resources").resolve(this.resourcePath)
     }
 
     override fun equals(other: Any?): Boolean {
@@ -34,16 +34,16 @@ class KotlinResourceDataSource(
         other as KotlinResourceDataSource
         if (sketch != other.sketch) return false
         if (request != other.request) return false
-        if (resourceName != other.resourceName) return false
+        if (resourcePath != other.resourcePath) return false
         return true
     }
 
     override fun hashCode(): Int {
         var result = sketch.hashCode()
         result = 31 * result + request.hashCode()
-        result = 31 * result + resourceName.hashCode()
+        result = 31 * result + resourcePath.hashCode()
         return result
     }
 
-    override fun toString(): String = "KotlinResourceDataSource('$resourceName')"
+    override fun toString(): String = "KotlinResourceDataSource('$resourcePath')"
 }